feat: successfully restructures schema+typeDefs, fixes input arg in regist function
