+++ apollo/src/typeDefs/auth/index.js
@@ -1,11 +1,15 @@
-import { inputDef } from './input'
-import { mutationDef } from './mutation'
-import { typeDef } from './type'
+import inputDef from './input'
+import mutationDef from './mutation'
+import typeDef from './type'
 
 
 
-export default {
-  inputDef,
+const auth = [inputDef,
   mutationDef,
   typeDef
-}
+]
+
+export default [inputDef,
+  mutationDef,
+  typeDef
+]
