+++ apollo/src/typeDefs/index.js
@@ -1,15 +1,15 @@
 //  TODO - Import Sub-TypeDefs
 import { query } from './query'
-import { enumDef } from './enum'
-import { user } from './user'
-import { auth } from './auth'
-import { stripe } from './stripe'
-import { chat } from './chat'
-import { listing } from './listing'
-
+import { mutation } from './mutation'
+import enumDef from './enum'
+import user from './user/'
+import auth from './auth'
+import stripe from './stripe'
+import chat from './chat'
+import listing from './listing'
 
 // TODO - Add Imported Sub-TypeDefs to array and export
 
-const typeDefs = [query, enumDef, user, auth, stripe, chat, listing]
+const schema = [query, mutation, enumDef, ...user, ...auth, ...stripe, ...chat, ...listing]
 
-export default typeDefs
+export default schema
