+++ apollo/src/typeDefs/mutation.js
@@ -1,0 +1,9 @@
+import { gql } from 'apollo-server'
+
+// ! Root mutation definition
+
+export const mutation = gql`
+  type Mutation {
+    root: String
+  }
+`
