+++ apollo/src/typeDefs/user/type.js
@@ -22,6 +22,4 @@ const typeDef = gql`
   }
 `
 
-export default {
-  typeDef
-}
+export default typeDef
